Work on improving readability of codebase
