+++ libs/backend/starrez/starrez-lib/StarRezClient.cs
@@ -1,11 +1,7 @@
 using System.Net.Http.Headers;
 using System.Text;
-using Microsoft.OpenApi.Models;
 using System.Net.Mime;
 using System.Xml;
-using Microsoft.OpenApi.Any;
-using Newtonsoft.Json.Linq;
-using Newtonsoft.Json;
 
 namespace StarRez;
 
@@ -65,540 +61,13 @@ public class StarRezClient
         return await response.Content.ReadAsStringAsync();
     }
 
-    /// <summary>
-    /// Adds development and production StarRez API servers to OpenApiDocument servers
-    /// </summary>
-    public void AddStarRezServers(OpenApiDocument document)
-    {
-        document.Servers = new List<OpenApiServer>();
-
-        foreach (var url in this.starrezApiUrls)
-        {
-            document.Servers.Add(new OpenApiServer()
-            {
-                Description = url.Key,
-                Url = url.Value
-            });
-        }
-    }
-
-    /// <summary>
-    /// Adds path parameters that were missing from original Swagger documentation
-    /// </summary>
-    private void _FixMissingPathParameters(OpenApiDocument document)
-    {
-        var missingParamPaths = document.Paths.Keys.Where(apiPath => apiPath.Contains("getreport") || apiPath.Contains("query")).ToList();
-        foreach (var apiPath in missingParamPaths)
-        {
-            if (apiPath.Contains("getreport"))
-            {
-                var updatedName = $"{apiPath}.{{format}}";
-                document.Paths.Add(updatedName, document.Paths[apiPath]);
-                document.Paths.Remove(apiPath);
-
-                foreach (var operation in document.Paths[updatedName].Operations)
-                {
-                    operation.Value.Parameters.Add(new OpenApiParameter()
-                    {
-                        Name = "format",
-                        In = ParameterLocation.Path,
-                        Required = true,
-                        Schema = new OpenApiSchema()
-                        {
-                            Type = "string"
-                        }
-                    });
-                }
-            }
-            else if (apiPath.Contains("query"))
-            {
-                document.Paths[apiPath].Operations[OperationType.Post].Description = "Allows the user to select ad-hoc data from the database using StarRez Query Language (StarQL). The data will be returned in a format appropriate for the desired content accept type.";
-                var baseOperationConfig = document.Paths[apiPath].Operations[OperationType.Post];
-                document.Paths[apiPath].Operations[OperationType.Post] = new OpenApiOperation(baseOperationConfig)
-                {
-                    RequestBody = new OpenApiRequestBody()
-                    {
-                        Description = "StarQL query to execute",
-                        Content = new Dictionary<string, OpenApiMediaType> { { "text/plain", new OpenApiMediaType() } },
-                        Required = true
-                    }
-                };
-
-                document.Paths[apiPath].Operations[OperationType.Get] = new OpenApiOperation(baseOperationConfig)
-                {
-                    Parameters = [new OpenApiParameter()
-                    {
-                        Name = "q",
-                        In = ParameterLocation.Query,
-                        Required = true,
-                        Schema = new OpenApiSchema()
-                        {
-                            Type = "string",
-                            Description = "StarQL query to execute"
-                        }
-                    }]
-                };
-            }
-        }
-    }
-
-    /// <summary>
-    /// Sets default `Accept` header to specify JSON as default response type
-    /// </summary>
-    private void _AddAcceptJsonHeader(OpenApiDocument document)
-    {
-        var pathsToUpdate = document.Paths.Keys.Where(apiPath => !apiPath.Contains("{format}")).ToList();
-
-        foreach (var apiPath in pathsToUpdate)
-        {
-            foreach (var operation in document.Paths[apiPath].Operations.Keys)
-            {
-                document.Paths[apiPath].Operations[operation].Parameters.Add(new OpenApiParameter()
-                {
-                    Name = "Accept",
-                    In = ParameterLocation.Header,
-                    Schema = new OpenApiSchema()
-                    {
-                        Type = "string",
-                        Default = new Microsoft.OpenApi.Any.OpenApiString("application/json"),
-                        Description = "Specifies the response return type following the MIME standard as specified by RFC 6838, section 4, for example `application/json` (for dynamic response typing, use `*/*`)",
-                    }
-                });
-            }
-        }
-    }
-
-    /// <summary>
-    /// Updates paths to fix improper path variable naming
-    /// </summary>
-    private void _UpdatePaths(OpenApiDocument document)
-    {
-        this._FixMissingPathParameters(document);
-        var pathsToUpdate = document.Paths.Keys.Where(apiPath => apiPath.Contains("tablename")).ToList();
-
-        foreach (var apiPath in pathsToUpdate)
-        {
-            var updatedName = apiPath.Replace("tablename", "tableName");
-            document.Paths.Add(updatedName, document.Paths[apiPath]);
-            document.Paths.Remove(apiPath);
-        }
-
-        this._AddAcceptJsonHeader(document);
-    }
-
-
-    /// <summary>
-    /// Replaces default HTTP methods with corrected methods
-    /// </summary>
-    public void CorrectHttpMethods(OpenApiDocument document)
-    {
-        foreach (var path in document.Paths)
-        {
-            var operationData = path.Value.Operations[OperationType.Post];
-
-            if ((path.Key.Contains("databaseinfo") ||
-                    path.Key.Contains("attachment") ||
-                    path.Key.Contains("select") ||
-                    (path.Key.Contains("photo") &&
-                        !path.Key.Contains("set")) ||
-                    path.Key.Contains("test") ||
-                    path.Key.Contains("get")) &&
-                    path.Value.Operations.All(operation => operation.Value.RequestBody == null))
-            {
-                path.Value.Operations.Add(OperationType.Get, operationData);
-                path.Value.Operations.Remove(OperationType.Post);
-            }
-            else if (path.Key.Contains("delete"))
-            {
-                path.Value.Operations.Add(OperationType.Delete, operationData);
-                path.Value.Operations.Remove(OperationType.Post);
-            }
-            else if (path.Key.Contains("update") &&
-                    !path.Key.Contains("post"))
-            {
-                path.Value.Operations.Add(OperationType.Patch, operationData);
-                path.Value.Operations.Add(OperationType.Put, operationData);
-                path.Value.Operations.Remove(OperationType.Post);
-            }
-            else if (path.Key.Contains("query"))
-            {
-                path.Value.Operations.Add(OperationType.Get, operationData);
-            }
-        }
-
-        this._UpdatePaths(document);
-    }
-
-    private Dictionary<string, OpenApiMediaType> _ConstructErrorResponses()
-    {
-        var errorResponses = new Dictionary<string, OpenApiMediaType>();
-        errorResponses.Add("application/json", new OpenApiMediaType()
-        {
-            Schema = new OpenApiSchema()
-            {
-                Description = "Response returned from StarRez API",
-                Type = "array",
-                Nullable = true,
-                Items = new OpenApiSchema()
-                {
-                    Description = "StarRez HTTP error message",
-                    Type = "object",
-                    Properties = new Dictionary<string, OpenApiSchema> {
-                        {
-                            "description",
-                            new OpenApiSchema() {
-                                Description = "Description of error",
-                                Type = "string"
-                            }
-                        }
-                    }
-                }
-            }
-        });
-        return errorResponses;
-    }
-
-    private Dictionary<string, OpenApiMediaType> _GetProperResponseType(string apiPath)
-    {
-        var responses = new Dictionary<string, OpenApiMediaType>();
-
-        if (apiPath.Contains("query"))
-        {
-            responses.Add("application/json", new OpenApiMediaType()
-            {
-                Schema = new OpenApiSchema()
-                {
-                    Description = "StarQL response data",
-                    Type = "object"
-                }
-            });
-        }
-
-        return responses;
-    }
-
-    private void _UpdateResponseCodes(OpenApiOperation operation, string apiPath)
-    {
-        operation.Responses = new OpenApiResponses();
-
-        operation.Responses.Add("200", new OpenApiResponse()
-        {
-            Description = "Everything went fine.",
-            Content = this._GetProperResponseType(apiPath)
-        });
-        operation.Responses.Add("400", new OpenApiResponse()
-        {
-            Description = "Your request is invalid, or badly formed, and we'll return an error message that tells you why.",
-            Content = this._ConstructErrorResponses()
-        });
-        operation.Responses.Add("403", new OpenApiResponse()
-        {
-            Description = "Your request is valid, but you do not have permission to select data from the specified table, or update the specified field.",
-            Content = this._ConstructErrorResponses()
-        });
-        operation.Responses.Add("404", new OpenApiResponse()
-        {
-            Description = "Your request is valid, but no data was found, or the table you are trying to use does not exist.",
-            Content = this._ConstructErrorResponses()
-        });
-    }
-
-    /// <summary>
-    /// Adds enum values to parameter schemas to improve documentation quality
-    /// </summary>
-    public void AddParameterEnums(OpenApiDocument document)
-    {
-        foreach (var path in document.Paths)
-        {
-            foreach (var operation in path.Value.Operations)
-            {
-                this._UpdateResponseCodes(operation.Value, path.Key);
-                for (int i = 0; i < operation.Value.Parameters.Count; i++)
-                {
-                    var parameterData = document
-                        .Paths[path.Key]
-                        .Operations[operation.Key]
-                        .Parameters[i];
-
-                    switch (operation.Value.Parameters[i].Name)
-                    {
-                        case "format":
-                            {
-                                IOpenApiAny[] formatList = [
-                                    new OpenApiString("atom"),
-                                    new OpenApiString("csv"),
-                                    new OpenApiString("htm"),
-                                    new OpenApiString("html"),
-                                    new OpenApiString("html-xml"),
-                                    new OpenApiString("json"),
-                                    new OpenApiString("xml")];
-
-                                parameterData
-                                    .Schema
-                                    .Enum = formatList.ToList();
-                                parameterData
-                                    .Schema
-                                    .Default = new OpenApiString("xml");
-                                break;
-                            }
-                    }
-                }
-            }
-        }
-    }
-
-    public void GenerateSpecialPropertyDescription(OpenApiSchema propertySchema, string modelName, string propertyName)
-    {
-        if (propertyName == "TableID")
-        {
-            propertySchema.Description = "ID of element in specified table";
-        }
-        else if (propertyName == "TableName")
-        {
-            propertySchema.Description = "Name of table to be used for reference";
-        }
-        else if (propertyName.Length != 3 &&
-                !propertyName.Contains("GUID") &&
-                propertyName.Contains("ID") &&
-                modelName != propertyName.Replace("ID", ""))
-        {
-            propertySchema.Description = $"References {propertyName.Substring(propertyName.IndexOf('_') + 1).Replace("ID", "")} table";
-        }
-        else if (propertyName.Length != 3 &&
-                !propertyName.Contains("GUID") &&
-                propertyName.Contains("ID"))
-        {
-            propertySchema.Description = $"Primary identification key for {propertyName.Substring(propertyName.IndexOf('_') + 1).Replace("ID", "")} table";
-        }
-    }
-
-    /// <summary>
-    /// Updates property data on models to improve quality of documentation
-    /// </summary>
-    public void ImproveModelProperties(OpenApiSchema propertySchema, string propertyName, string existingType)
-    {
-        switch (existingType)
-        {
-            case string dateType when dateType.Contains("datetime"):
-                {
-                    propertySchema.Type = "string";
-                    propertySchema.Format = "date-time";
-                    propertySchema.Description = "Date and time as defined by RFC 3339, section 5.6, for example, 2017-07-21T17:32:28Z";
-                    break;
-                }
-            case "date":
-                {
-                    propertySchema.Type = "string";
-                    propertySchema.Format = "date";
-                    propertySchema.Description = "Date as defined by RFC 3339, section 5.6, for example, 2017-07-21";
-                    break;
-                }
-            case "timestamp":
-                {
-                    propertySchema.Type = "string";
-                    propertySchema.Format = "time";
-                    propertySchema.Description = "Time as defined by RFC 3339, section 5.6, for example, 17:32:28Z";
-                    break;
-                }
-            case "money":
-                {
-                    propertySchema.Type = "number";
-                    propertySchema.Format = "decimal";
-                    propertySchema.Description = "A fixed point decimal number of unspecified precision and range";
-                    break;
-                }
-            case "decimal":
-                {
-                    goto case "money";
-                }
-            case "bigdecimal":
-                {
-                    goto case "money";
-                }
-            case "binary":
-                {
-                    propertySchema.Type = "string";
-                    propertySchema.Format = "binary";
-                    propertySchema.Description = "Any sequence of octets";
-                    break;
-                }
-            case "byte":
-                {
-                    propertySchema.Type = "string";
-                    propertySchema.Format = "byte";
-                    propertySchema.Description = "Base64 encoded data as defined by RFC4648, section 6";
-                    break;
-                }
-            case "guid":
-                {
-                    propertySchema.Type = "string";
-                    propertySchema.Format = "guid";
-                    propertySchema.Description = "A Globally Unique Identifier, defined as UUID by RFC 9562, section 4, for example, f81d4fae-7dec-11d0-a765-00a0c91e6bf6";
-                    break;
-                }
-            case "longstring":
-                {
-                    goto case "binary";
-                }
-            case "short":
-                {
-                    propertySchema.Type = "integer";
-                    propertySchema.Format = "uint16";
-                    propertySchema.Description = "Unsigned 16-bit integer";
-                    break;
-                }
-            default:
-                {
-                    propertySchema.Type = existingType;
-                    break;
-                }
-        }
-
-
-        if (propertyName.Contains("Email") || propertyName == "PortalAuthProviderUserID")
-        {
-            propertySchema.Type = "string";
-            propertySchema.Format = "email";
-            propertySchema.Description = "An email address, defined as Mailbox by RFC5321, section 2.3.11, for example, example@domain.com";
-        }
-    }
-
-    private async Task<string> _GetStarRezModelDefinitions(bool? dev)
+    public async Task<string> GetStarRezModelDefinitions(bool? dev)
     {
         var request = new HttpRequestMessage(HttpMethod.Get, $"{apiUrl}/starrez/models");
         request.Headers.Add("dev", (dev ?? false).ToString());
         var response = await this.client.SendAsync(request);
         response.EnsureSuccessStatusCode();
         return await response.Content.ReadAsStringAsync();
-    }
-
-    /// <summary>
-    /// Ensures that all schema properties are properly set (fixes problems with JSON deserialization)
-    /// </summary>
-    private void _FixSchemaFormatting(JObject schemaValue, OpenApiSchema schema)
-    {
-        if (schemaValue.TryGetValue("type", out var type))
-        {
-            schema.Type = type.Value<string>();
-        }
-
-        if (schemaValue.TryGetValue("format", out var format))
-        {
-            schema.Format = format.Value<string>();
-        }
-
-        if (schemaValue.TryGetValue("maxLength", out var maxLength))
-        {
-            schema.MaxLength = maxLength.Value<int>();
-        }
-
-        if (schemaValue.TryGetValue("nullable", out var nullable))
-        {
-            schema.Nullable = nullable.Value<bool>();
-        }
-
-        if (schemaValue.TryGetValue("description", out var description))
-        {
-            schema.Description = description.Value<string>();
-        }
-
-        if (schemaValue.TryGetValue("required", out var required) && required is JArray)
-        {
-            schema.Required = ((JArray)required).Select<JToken, string>(item =>
-                item.Value<string>() ?? ""
-            ).Where(s => !string.IsNullOrWhiteSpace(s)).ToHashSet();
-        }
-
-        if (schemaValue.TryGetValue("enum", out var enumToken) &&
-enumToken is JArray)
-        {
-            schema.Enum = ((JArray)enumToken).Select<JToken, IOpenApiAny>(item =>
-             item.Type switch
-                {
-                    JTokenType.Integer => new OpenApiInteger((int)item.Value<long>()),
-                    JTokenType.String => new OpenApiString(item.Value<string>()),
-                    _ => new OpenApiString(item.ToString())
-                }).Cast<IOpenApiAny>().ToList();
-        }
-
-        if (schemaValue.TryGetValue("oneOf", out var typesToken) &&
-        typesToken is JArray)
-        {
-            schema.OneOf = ((JArray)typesToken)
-                .Select(item =>
-                        new OpenApiSchema()
-                        {
-                            Type = item["type"]?.Value<string>() ?? ""
-                        })
-            .ToList();
-        }
-    }
-
-    /// <summary>
-    /// Makes an HTTP request to get all StarRez model definitions
-    /// </summary>
-    public async Task GetStarRezModels(bool? dev, OpenApiDocument document)
-    {
-        var models = JsonConvert.DeserializeObject<Dictionary<string, JObject>>(
-                await this._GetStarRezModelDefinitions(dev));
-
-        var formattedModels = new Dictionary<string, OpenApiSchema>();
-
-        foreach (var model in models ?? [])
-        {
-            var modelName = model.Key;
-            var schema = new OpenApiSchema();
-
-            this._FixSchemaFormatting(model.Value, schema);
-
-            if (model.Value.TryGetValue("properties", out var modelProperties) &&
-                modelProperties is JObject properties)
-            {
-                foreach (var property in properties)
-                {
-                    var propertyName = property.Key;
-                    var propertySchema = new OpenApiSchema();
-
-                    this._FixSchemaFormatting(
-                        property.Value as JObject ?? new JObject(), propertySchema);
-
-                    schema.Properties[propertyName] = propertySchema;
-                }
-            }
-
-            formattedModels[modelName] = schema;
-        }
-
-        formattedModels = formattedModels.OrderBy(key => key.Key).ToDictionary();
-
-        document.Components = new OpenApiComponents()
-        {
-            Schemas = formattedModels,
-        };
-    }
-
-    public void AddStarRezAuth(OpenApiDocument document)
-    {
-        var basicAuth = new OpenApiSecurityScheme
-        {
-            Description = "Basic authentication scheme",
-            Reference = new OpenApiReference
-            {
-                Type = ReferenceType.SecurityScheme,
-                Id = "Basic"
-            },
-            Scheme = "Basic",
-            Name = "Authorization",
-            Type = SecuritySchemeType.Http,
-            In = ParameterLocation.Header,
-        };
-
-        document.Components.SecuritySchemes.Add(basicAuth.Scheme, basicAuth);
-        document.SecurityRequirements.Add(new OpenApiSecurityRequirement {
-              {
-                  basicAuth,
-                  new List<string>()
-              }});
     }
 
     /// <summary>
